feat: add registry port to config; use injection over global var
This adds quite a bit of new features. It adds logic for passing a 'registryPort' arg to force use a port. This still needs some adjustments, but the key point is that before, this variable was globally set. Now, it's encapsulated and injected into the validadtor, ensuring port fallbacks stay in sync between the parent LT and child Validator...plus Registry.